requirements.txt
numpy 
opencv-python>=3.4.2
opencv-contrib-python >=3.4.2
imutils
Pillow
tkinter
pyttk